+++ src/nodes/Conditional.ts
@@ -88,8 +88,7 @@ export default class Conditional extends Expression {
             {
                 name: 'condition',
                 kind: node(Expression),
-                label: (locales: Locales) =>
-                    locales.get((l) => l.node.Conditional.condition),
+                label: () => (l) => l.node.Conditional.condition,
                 // Must be boolean typed
                 getType: () => BooleanType.make(),
             },
@@ -97,16 +96,14 @@ export default class Conditional extends Expression {
             {
                 name: 'yes',
                 kind: node(Expression),
-                label: (locales: Locales) =>
-                    locales.get((l) => l.node.Conditional.yes),
+                label: () => (l) => l.node.Conditional.yes,
                 space: true,
                 indent: true,
             },
             {
                 name: 'no',
                 kind: node(Expression),
-                label: (locales: Locales) =>
-                    locales.get((l) => l.node.Conditional.no),
+                label: () => (l) => l.node.Conditional.no,
                 space: true,
                 indent: true,
             },