+++ src/nodes/Node.ts
@@ -2,7 +2,7 @@
 /* eslint-disable @typescript-eslint/no-explicit-any */
 /* eslint-disable @typescript-eslint/no-unused-vars */
 import type Conflict from '@conflicts/Conflict';
-import type { DocText, LocaleText, Template } from '@locale/LocaleText';
+import type { DocText, LocaleText } from '@locale/LocaleText';
 import type {
     DescriptiveNodeText,
     NodeDescriptor,
@@ -12,7 +12,7 @@ import type Spaces from '@parser/Spaces';
 import type { BasisTypeName } from '../basis/BasisConstants';
 import type Purpose from '../concepts/Purpose';
 import type Locales from '../locale/Locales';
-import type { TemplateInput } from '../locale/Locales';
+import type { LocaleTextAccessor, TemplateInput } from '../locale/Locales';
 import type BasisCharacter from '../lore/BasisCharacter';
 import type Context from './Context';
 import type Definition from './Definition';
@@ -660,7 +660,7 @@ export default abstract class Node {
         locales: Locales,
         context: Context,
         root: Root,
-    ): Template | undefined {
+    ): LocaleTextAccessor | undefined {
         const label = this.getFieldOfChild(child)?.label;
         return label ? label(locales, child, context, root) : undefined;
     }
@@ -692,7 +692,7 @@ export type Field = {
         child: Node,
         context: Context,
         root: Root,
-    ) => Template;
+    ) => LocaleTextAccessor;
     /** True if a preceding space is preferred the node */
     space?: boolean | ((node: Node) => boolean);
     /** True if the field should be indented if on a new line */