+++ src/nodes/TableLiteral.ts
@@ -198,7 +198,7 @@ export default class TableLiteral extends Expression {
             {
                 name: 'type',
                 kind: node(TableType),
-                label: (locales: Locales) => locales.get((l) => l.term.table),
+                label: () => (l) => l.term.table,
             },
             { name: 'rows', kind: list(true, node(Row)), newline: true },
         ];