+++ src/nodes/TypePlaceholder.ts
@@ -3,7 +3,6 @@ import Placeholder from '@conflicts/Placeholder';
 import type LocaleText from '@locale/LocaleText';
 import type { NodeDescriptor } from '@locale/NodeTexts';
 import type { BasisTypeName } from '../basis/BasisConstants';
-import type Locales from '../locale/Locales';
 import Characters from '../lore/BasisCharacters';
 import { node, type Grammar, type Replacement } from './Node';
 import PlaceholderToken from './PlaceholderToken';
@@ -43,7 +42,7 @@ export default class TypePlaceholder extends Type {
             {
                 name: 'placeholder',
                 kind: node(Sym.Placeholder),
-                label: (locales: Locales) => locales.get((l) => l.term.type),
+                label: () => (l) => l.term.type,
             },
         ];
     }