+++ CHANGELOG.md
@@ -20,6 +20,7 @@ Dates are in `YYYY-MM-DD` format and versions are in [semantic versioning](http:
 - Fixed #528: Deploy to prod on push or merge to main.
 - Refactored widgets to use locale accessor functions instead of strings, for provenance.
 - Refactored nodes to return locale accessor function instead of specific locale, for provenance.
+- Refactored node labels to return locale accessor function, for provenance.
 
 ## 0.16.0 2025-02-25